refactor: decouple package inputs
