+++ src/inputs/getters.ts
@@ -1,6 +1,5 @@
 import type { SupportedSvgoVersions } from "../svgo";
-import type { error, Inputter, InputterOptions } from "../types";
-import type { InputValue } from "./types";
+import type { error, Inputter, InputterOptions, InputValue } from "./types";
 
 import errors from "../errors";
 import {
@@ -155,7 +154,10 @@ function getSvgoVersion(
     ];
   }
 
-  return [input as InputValue<SupportedSvgoVersions>, null];
+  return [
+    input as InputValue<SupportedSvgoVersions>,
+    null,
+  ];
 }
 
 export {
